# tiltfile: parallelize docker-compose config calls for faster manifest load [ch1201]