this will save us .4s per dc.yml when loading manifests.

in particular, a big gain for tests:
```
$ # on master
$ go test github.com/windmilleng/tilt/internal/engine -count=1 && go test github.com/windmilleng/tilt/internal/tiltfile -count=1
ok  	github.com/windmilleng/tilt/internal/engine	15.682s
ok  	github.com/windmilleng/tilt/internal/tiltfile	14.437s

$ # on this branch
$ go test github.com/windmilleng/tilt/internal/engine -count=1 && go test github.com/windmilleng/tilt/internal/tiltfile -count=1
ok  	github.com/windmilleng/tilt/internal/engine	10.958s
ok  	github.com/windmilleng/tilt/internal/tiltfile	8.815s
```
So that's a savings of c. 10s per test suite run.